set up initial game.jsx.  renamed old game.jsx to _game.jsx
