+++ client/src/Profile.jsx
@@ -35,7 +35,7 @@ const Profile = () => {
   }, [])
 
   const getGames = async () => {
-    const response = await fetch(`http://localhost:5432/api/games/?id=${user.id}`, {
+    const response = await fetch(`http://localhost:3001/api/games/?id=${user.id}`, {
       method: 'GET',
       headers: {
         'Content-Type': 'application/json',
